+++ docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased]
+## [Unreleased 1.0.1]
+
+- Added support for USB Device Id as serial port to motor_config.py
 
 ## 1.0.0